Implement global registration algorithm and tasks
This commit adds the implementation of a global registration algorithm for roughly aligning two point clouds. It also includes tasks to fit 'r1.pcd' and 'r2.pcd', as well as 'car1.ply' and 'car2.ply' from the 'global_registration' folder. Additionally, it challenges the implementation of local registration and attempts to reconstruct the car from the images in the 'car_challenge' folder.